Fix scroller_barWidth issue on iOS
This change fixes an issue where the last column header gets cut off on MobileSafari (possibly other mobile browsers?). 

MobileSafari doesn't allocate horizontal space for the vertical scollbar of vertically-scrolling regions, so the getBarWidth() function correctly returns 0 on iOS. The previous expression of (tsScroller.getBarWidth() || 15) incorrectly evaluates that 0 as "falsy" and sets scroller_barSetWidth to 15, which cuts off the last column header. This change explicitly tests for a null result from tsScroller.getBarWidth() and only applies 15px in that case 

I also expanded that section of code a little bit and modified the comment to make it more clear what's going on